+++ codec.go
@@ -24,6 +24,7 @@ type Reader interface {
 
 // Writer is used to marshal pkg and write to session
 type Writer interface {
+	// if @Session is udpGettySession, the second parameter is UDPContext.
 	Write(Session, interface{}) error
 }